requirements.txt
streamlit
pandas
google-generativeai
spacy
altair
st-gsheets-connection